requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile requirements.in
#
absl-py==2.1.0
    # via
    #   chex
    #   optax
    #   orbax-checkpoint
aiohttp==3.9.1
    # via
    #   datasets
    #   fsspec
aiosignal==1.3.1
    # via aiohttp
async-timeout==4.0.3
    # via aiohttp
attrs==23.2.0
    # via aiohttp
black==23.12.1
    # via -r requirements.in
certifi==2023.11.17
    # via requests
charset-normalizer==3.3.2
    # via requests
chex==0.1.85
    # via optax
click==8.1.7
    # via black
contourpy==1.2.0
    # via matplotlib
cycler==0.12.1
    # via matplotlib
datasets==2.16.1
    # via -r requirements.in
dill==0.3.7
    # via
    #   datasets
    #   multiprocess
etils[epath,epy]==1.5.2
    # via orbax-checkpoint
filelock==3.13.1
    # via
    #   datasets
    #   huggingface-hub
    #   torch
    #   transformers
flax==0.7.5
    # via -r requirements.in
fonttools==4.47.2
    # via matplotlib
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
fsspec[http]==2023.10.0
    # via
    #   datasets
    #   etils
    #   huggingface-hub
    #   torch
huggingface-hub==0.20.2
    # via
    #   datasets
    #   tokenizers
    #   transformers
idna==3.6
    # via
    #   requests
    #   yarl
importlib-metadata==7.0.1
    # via jax
importlib-resources==6.1.1
    # via
    #   etils
    #   matplotlib
jax==0.4.23
    # via
    #   -r requirements.in
    #   chex
    #   flax
    #   optax
    #   orbax-checkpoint
jaxlib==0.4.23
    # via
    #   -r requirements.in
    #   chex
    #   optax
    #   orbax-checkpoint
jinja2==3.1.3
    # via torch
kiwisolver==1.4.5
    # via matplotlib
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.3
    # via jinja2
matplotlib==3.8.2
    # via -r requirements.in
mdurl==0.1.2
    # via markdown-it-py
ml-dtypes==0.3.2
    # via
    #   jax
    #   jaxlib
    #   tensorstore
mpmath==1.3.0
    # via sympy
msgpack==1.0.7
    # via
    #   flax
    #   orbax-checkpoint
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
multiprocess==0.70.15
    # via datasets
mypy==1.8.0
    # via -r requirements.in
mypy-extensions==1.0.0
    # via
    #   black
    #   mypy
nest-asyncio==1.5.9
    # via orbax-checkpoint
networkx==3.2.1
    # via torch
numpy==1.26.3
    # via
    #   -r requirements.in
    #   chex
    #   contourpy
    #   datasets
    #   flax
    #   jax
    #   jaxlib
    #   matplotlib
    #   ml-dtypes
    #   opt-einsum
    #   optax
    #   orbax-checkpoint
    #   pandas
    #   pyarrow
    #   scipy
    #   tensorstore
    #   transformers
opt-einsum==3.3.0
    # via jax
optax==0.1.8
    # via flax
orbax==0.1.9
    # via -r requirements.in
orbax-checkpoint==0.5.0
    # via
    #   flax
    #   orbax
packaging==23.2
    # via
    #   black
    #   datasets
    #   huggingface-hub
    #   matplotlib
    #   transformers
pandas==2.1.4
    # via datasets
pathspec==0.12.1
    # via black
pillow==10.2.0
    # via matplotlib
platformdirs==4.1.0
    # via black
protobuf==4.25.2
    # via orbax-checkpoint
pyarrow==14.0.2
    # via datasets
pyarrow-hotfix==0.6
    # via datasets
pygments==2.17.2
    # via rich
pyparsing==3.1.1
    # via matplotlib
python-dateutil==2.8.2
    # via
    #   matplotlib
    #   pandas
pytz==2023.3.post1
    # via pandas
pyyaml==6.0.1
    # via
    #   datasets
    #   flax
    #   huggingface-hub
    #   orbax-checkpoint
    #   transformers
regex==2023.12.25
    # via transformers
requests==2.31.0
    # via
    #   datasets
    #   fsspec
    #   huggingface-hub
    #   transformers
rich==13.7.0
    # via flax
safetensors==0.4.1
    # via transformers
scipy==1.11.4
    # via
    #   jax
    #   jaxlib
six==1.16.0
    # via python-dateutil
sympy==1.12
    # via torch
tensorstore==0.1.52
    # via
    #   flax
    #   orbax-checkpoint
tokenizers==0.15.0
    # via transformers
tomli==2.0.1
    # via
    #   black
    #   mypy
toolz==0.12.0
    # via chex
torch==2.1.2
    # via -r requirements.in
tqdm==4.66.1
    # via
    #   -r requirements.in
    #   datasets
    #   huggingface-hub
    #   transformers
transformers==4.36.2
    # via -r requirements.in
types-tqdm==4.66.0.20240106
    # via -r requirements.in
typing-extensions==4.9.0
    # via
    #   black
    #   chex
    #   etils
    #   flax
    #   huggingface-hub
    #   mypy
    #   orbax-checkpoint
    #   torch
tzdata==2023.4
    # via pandas
urllib3==2.1.0
    # via requests
xxhash==3.4.1
    # via datasets
yarl==1.9.4
    # via aiohttp
zipp==3.17.0
    # via
    #   etils
    #   importlib-metadata
    #   importlib-resources
